Introduce Database Command Line Interface
- Add basic support for CLI with a sole `create` command which creates
  an empty database. More commands will be added in the future
- Add script which performs CLI regression tests
- Modify travis pipeline to build library in release mode also in order
  to run CLI regression tests.
- Refactor `Database::create` to throw an error in case another database
  already resides inside a specific directory

+++ db/src/io.rs
@@ -49,9 +49,15 @@ impl Io {
                 "Database name contains forbidden characters",
             ));
         }
-
         // Create an empty metadata structure inside a specified directory
         let database_path = Path::new(&path).canonicalize()?.join(name);
+        if database_path.exists() {
+            return Err(Error::custom_err(
+                CustomKind::InvalidArgument,
+                "Database already exists",
+            ));
+        }
+
         let metadata_path = database_path.join(Self::METADATA_DIR);
         fs::create_dir_all(&metadata_path)?;
         File::create(metadata_path.join(Self::METADATA_FILE))?;
